Typo fix. Also read extdesc.
Remove the import for this from the CSS if this proves too bothersome to use. Before I was using the notifications space but that was making the announcement box be on top of other things, which was really, really dumb. I may revert back to that and just give everything else a higher z-index, but only if I have to.