Release 04/09/15.5
Removed vote:before from the other colours section because it doesn't need to exist. I keep on forgetting to DO THIS but now it's done and out of the way.

Also resolves the issue when moving back and forth between the socmed buttons; The avatar at the DJ booth would be masked with a black box.